Add logic to detect spurious timer interrupt
I previously thought I was getting a double fault exception with no
error code on the stack, but it was a spurious timer interrupt after
re-enabling interrupts in protected mode, so this change modifies the
interrupt handler for vector 8 to look at the stack to see if topmost
DWORD is 0, which would indicate a double fault, and pop it if so. The
change also uses a multi-line macro to define the interrupt handlers.